chore: fixed issues

+++ contracts/src/interfaces/components/IOracleManager.1.sol
@@ -158,7 +158,7 @@ interface IOracleManagerV1 {
         uint256 validatorsSkimmedBalance;
         uint256 validatorsExitedBalance;
         uint256 validatorsExitingBalance;
-        uint256 validatorsCount;
+        uint32 validatorsCount;
         bool rebalanceDepositToRedeemMode;
         bool slashingContainmentMode;
     }
